% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_incidence_map.R
\name{mod_fktselect_server}
\alias{mod_fktselect_server}
\title{updates the list of fokontany to choose from}
\usage{
mod_fktselect_server(id)
}
\description{
updates the list of fokontany to choose from
}
