% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_incidence-timeseries.R
\name{plot_inc_time}
\alias{plot_inc_time}
\title{Plot incidence timeseries}
\usage{
plot_inc_time(
  historical = T,
  current.month = as.Date("2020-12-01"),
  communeSelect = "District",
  fktSelect = "Selectionner"
)
}
\arguments{
\item{plot_data}{dataframe to plot}

\item{color.scale}{color for lines used in plot}

\item{size}{vector of linewidths used in plot}
}
\description{
Plot incidence timeseries
}
