% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotlyTime.R
\name{plotly_timeseries}
\alias{plotly_timeseries}
\title{Create a plotly plot}
\usage{
plotly_timeseries(input_data, ci_data, ylabel, ptitle)
}
\arguments{
\item{input_data}{dataframe to use to plot all data}

\item{ci_data}{dataframe of CIs for error in predictions}

\item{ylabel}{label for y-axis of plot}

\item{ptitle}{label for plot}
}
\description{
Creates plotly plots of time series
}
