% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotlyTime.R
\name{vline}
\alias{vline}
\title{Create a vertical line in plotly}
\usage{
vline(x = 0, color = "gray50")
}
\arguments{
\item{x}{x-intercept of line}

\item{color}{color of the line}
}
\description{
function to create a vertical line in plotly
}
