% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_incidence_time.R
\name{plot_inc_time}
\alias{plot_inc_time}
\title{Plot incidence timeseries}
\usage{
plot_inc_time(
  historical = T,
  current.month = as.Date("2020-12-01"),
  communeSelect = "District",
  fktSelect = "Selectionner"
)
}
\arguments{
\item{current.month}{first day of current month, used to draw prediction line}

\item{communeSelect}{name of commune selected in UI}

\item{fktSelect}{name of fokontany selected in UI}

\item{plot_data}{dataframe to plot}
}
\description{
Plot incidence timeseries
}
